refactor: based on PR review
